removing the check in fitness() for input len(string) == len(TARGET). Added a comment saying that the lengths should match
I didn't want this check in the doctests anymore (it seemed to be cluttering), and I don't think the multi line check at the start of the function was necessary either. If this fails it should be clear what the problem is. Besides, we're creating our own chromos, so there shouldn't be any situation that they would not be same length as target.